Error Messages May Now Include Fields
This patch enhances Goof with the flags IncludeFieldsInError,
IncludeFieldsInFormat, and IncludeFieldsInString. These flags are
available globally and per-Error instance. When enabled, these flags
will include an error's fields inside of the emitted error message. This
will provide a central way to ensure that inner errors and error field
data is emitted to projects that don't specifically extract it via other
means.